% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeffery.R
\name{sim_data}
\alias{sim_data}
\title{Simulate dataset for testing three bias-corrected meethods}
\usage{
sim_data(theta, n, dose_set)
}
\arguments{
\item{theta}{True parameters of Emax model. The order of the variables is (E0,log(ED50),Emax).}

\item{n}{number of observations.}

\item{dose_set}{A vector indicate the dose set for the dose-response relationship.}
}
\value{
A dataframe of simulated dataset.
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 theta_true=matrix(c(qlogis(0.1),qlogis(0.8)-qlogis(0.1),log(7.5)),1,3)
colnames(theta_true)<- c('e_0','emax','led_50')
theta_true <- as.data.frame(theta_true)
dose_set <- c(0,7.5,22.5,75,225)
n=355
data <-sim_data(theta_true,n,dose_set)
 }
}
}
